requirements.txt

fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.9
pydantic-settings>=2.6
SQLAlchemy>=2.0
psycopg[binary,pool]>=3.2
alembic>=1.13

python-dotenv>=1.0
loguru>=0.7
numpy>=2.1

# AI stack
langchain>=0.3.9
langgraph>=0.2.39
faiss-cpu>=1.8.0
openai>=1.51

# Optional: file parsing later
python-multipart>=0.0.9
